lista2/requirements.txt
rpyc~=5.0.1
pycryptodome==3.10.1
